feature(theme) :hammer: improve theme provider and change homepage
