+++ package.json
@@ -5,44 +5,46 @@
     "ios": "react-native run-ios",
     "web": "expo start --web",
     "start": "react-native start",
-    "test": "jest",
+    "test": "jest --passWithNoTests",
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "prepare": "husky install"
   },
   "dependencies": {
+    "@expo-google-fonts/montserrat": "^0.2.2",
     "@types/metro-config": "^0.66.0",
     "axios": "^0.26.1",
-    "expo": "~41.0.1",
-    "expo-splash-screen": "~0.10.2",
-    "expo-updates": "~0.5.4",
-    "react": "16.13.1",
-    "react-dom": "16.13.1",
-    "react-native": "~0.63.4",
-    "react-native-gesture-handler": "~1.10.2",
-    "react-native-reanimated": "~2.1.0",
-    "react-native-screens": "~3.0.0",
+    "expo": "^45.0.0",
+    "expo-splash-screen": "~0.15.1",
+    "expo-updates": "~0.13.1",
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "react-native": "0.68.2",
+    "react-native-gesture-handler": "~2.2.1",
+    "react-native-reanimated": "~2.8.0",
+    "react-native-screens": "~3.11.1",
+    "react-native-svg": "^12.3.0",
     "react-native-svg-transformer": "^1.0.0",
     "react-native-unimodules": "~0.13.3",
-    "react-native-web": "~0.13.12",
+    "react-native-web": "0.17.7",
     "styled-components": "^5.3.5"
   },
   "devDependencies": {
-    "@babel/core": "^7.9.0",
+    "@babel/core": "^7.12.9",
     "@eslint/create-config": "^0.1.2",
-    "@types/react": "~16.9.35",
-    "@types/react-native": "~0.63.2",
+    "@types/react": "~17.0.21",
+    "@types/react-native": "~0.67.6",
     "@types/styled-components": "^5.1.25",
     "@typescript-eslint/eslint-plugin": "^5.20.0",
     "@typescript-eslint/parser": "^5.20.0",
-    "babel-preset-expo": "~8.3.0",
+    "babel-preset-expo": "~9.1.0",
     "eslint": "^8.14.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.29.4",
     "husky": "^7.0.4",
-    "jest-expo": "~41.0.0",
+    "jest-expo": "^45.0.0",
     "prettier": "^2.6.2",
-    "typescript": "~4.0.0"
+    "typescript": "~4.3.5"
   },
   "jest": {
     "preset": "react-native"
