+++ src/views/theme/assets/fonts.tsx
@@ -1,0 +1,8 @@
+const fonts = {
+  default: {
+    family: 'Montserrat_400Regular',
+    size: '24px',
+  },
+};
+
+export default fonts;
