+++ src/views/theme/assets/icons/index.tsx
