+++ src/views/theme/themeProvider/index.tsx
@@ -1,20 +1,25 @@
 import React, { ReactNode } from 'react';
 import { ThemeProvider } from 'styled-components/native';
 
-import colors from './colors';
-import icons from './icons';
+import colors from '../assets/colors';
+import fonts from '../assets/fonts';
+import { SafeScreen } from './styles';
 
 interface ProviderProps {
   children: ReactNode;
 }
 
 const theme = {
-  icons: icons,
   colors: colors,
+  fonts: fonts,
 };
 
 function AppThemeProvider({ children }: ProviderProps) {
-  return <ThemeProvider theme={theme}>{children}</ThemeProvider>;
+  return (
+    <ThemeProvider theme={theme}>
+      <SafeScreen>{children}</SafeScreen>
+    </ThemeProvider>
+  );
 }
 
 export default AppThemeProvider;
